requirements.txt
dash>=2.18.1
dash-bootstrap-components>=1.6.0
Flask>=3.0.3
requests>=2.32.3
pytz>=2024.1
python-dotenv>=1.0.1